rename tableResponsive to pgResponsive
