+++ resources/js/components/index.js
@@ -1,14 +1,14 @@
 import pgToggleable from './pg-toggleable'
 import pgFlatpickr from "./pg-flatpickr";
 import pgEditable from "./pg-editable";
-import tableResponsive from './tableResponsive';
+import pgResponsive from './pgResponsive';
 import pgTomSelect from "./select/tomSelect";
 import pgSlimSelect from "./select/slimSelect";
 
 window.pgToggleable = pgToggleable
 window.pgFlatpickr = pgFlatpickr
 window.pgEditable = pgEditable
-window.tableResponsive = tableResponsive
+window.pgResponsive = pgResponsive
 window.pgTomSelect = pgTomSelect
 window.pgSlimSelect = pgSlimSelect
