+++ resources/js/components/pgResponsive/index.js
