+++ resources/js/components/pgResponsive/onResize.js
